requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
frozenlist==1.4.1
idna==3.8
magic-filter==1.0.12
Markdown==3.7
multidict==6.1.0
numpy==2.1.0
opencv-python==4.10.0.84
pillow==10.4.0
psycopg2==2.9.9
pydantic==2.9.2
pydantic_core==2.23.4
python-barcode==0.15.1
python-dotenv==1.0.1
redis==5.0.8
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.2
yarl==1.11.1
